+++ HISTORY.md
@@ -5,6 +5,9 @@
 - Removed `log.tqdm`
 - Added binning functions to `pelutils.ds.plot`
 - Added `pelutils.ds.distributions` with `scipy` distributions that use same notation as Jim Pitman's "Probability"
+- Added `examples` directory
+
+  Currently only has plotting examples
 
 ## 0.6.7